+++ index.js
@@ -21,6 +21,7 @@ app.use('/api/cars', require(path.join(__dirname, "/routes/api/create-cars-data.
 app.use('/api/subscribe', require(path.join(__dirname, "/routes/api/email-subscribe.js")));
 app.use('/api/account', require(path.join(__dirname, "/routes/api/account.js")));
 app.use('/api/booking', require(path.join(__dirname, "/routes/api/booking.js")));
+app.use('/api/invoice', require(path.join(__dirname, "/routes/api/invoice.js")));
 
 
 app.listen(port, () => {